Updating example of using pre-existing TF model
Adding MNIST image index as an argument so it's easier to try different images.

Fixing a bug with the image output printed to the terminal.

Adding output so it's clear which number the program guessed the image represents.

Note this script is still a WIP while I research ways to not have to specify every little detail about the variables in the graph.  There should be mechanisms to load variables and the graph defining their relationships, but getting it working is tricky.